requirements.txt
# Apache Airflow
apache-airflow==2.7.3
pendulum==2.1.2

# Data processing
numpy==1.24.3
pandas==2.0.3

# Machine Learning
tensorflow==2.13.0
keras==2.13.1
scikit-learn==1.3.0

# Visualization
matplotlib==3.7.2

# Email
sendgrid==6.10.0

# Database (PostgreSQL)
psycopg2-binary==2.9.9
SQLAlchemy==2.0.23

# Kafka (Real-time Streaming)
kafka-python==2.0.2

# Technical Analysis
ta==0.11.0

# SSL handling (built-in with Python)

